@@ -7,7 +7,7 @@ issue_tracker: https://github.com/nlfiedler/tanuki/issues
 publish_to: 'none'
 
 environment:
-  sdk: ">=2.12.0"
+  sdk: ">=3.0.0"
 
 dependencies:
   flutter:
