@@ -5,34 +5,42 @@ import 'package:bloc_test/bloc_test.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:oxidized/oxidized.dart';
-import 'package:tanuki/core/domain/entities/attributes.dart';
+import 'package:tanuki/core/domain/entities/asset.dart';
 import 'package:tanuki/core/domain/repositories/entity_repository.dart';
-import 'package:tanuki/core/domain/usecases/get_all_locations.dart';
+import 'package:tanuki/core/domain/usecases/get_asset_locations.dart';
 import 'package:tanuki/core/error/failures.dart';
-import 'package:tanuki/features/browse/preso/bloc/raw_locations_bloc.dart';
+import 'package:tanuki/features/import/preso/bloc/raw_locations_bloc.dart';
 
 class MockEntityRepository extends Mock implements EntityRepository {}
 
 void main() {
   late MockEntityRepository mockEntityRepository;
-  late GetAllLocations usecase;
+  late GetAssetLocations usecase;
 
   group('normal cases', () {
     const incoming = [
-      Location(label: 'tokyo', count: 806),
-      Location(label: 'paris', count: 269),
-      Location(label: 'london', count: 23),
+      AssetLocation(label: None(), city: Some('Tokyo'), region: Some('Japan')),
+      AssetLocation(
+        label: Some('eiffel'),
+        city: Some('Paris'),
+        region: Some('France'),
+      ),
+      AssetLocation(label: None(), city: Some('London'), region: Some('UK')),
     ];
     // the incoming list will be served in sorted order
     const expected = [
-      Location(label: 'london', count: 23),
-      Location(label: 'paris', count: 269),
-      Location(label: 'tokyo', count: 806),
+      AssetLocation(label: None(), city: Some('Tokyo'), region: Some('Japan')),
+      AssetLocation(
+        label: Some('eiffel'),
+        city: Some('Paris'),
+        region: Some('France'),
+      ),
+      AssetLocation(label: None(), city: Some('London'), region: Some('UK')),
     ];
     setUp(() {
       mockEntityRepository = MockEntityRepository();
-      usecase = GetAllLocations(mockEntityRepository);
-      when(() => mockEntityRepository.getAllLocations(true))
+      usecase = GetAssetLocations(mockEntityRepository);
+      when(() => mockEntityRepository.getAssetLocations())
           .thenAnswer((_) async => Ok(List.from(incoming)));
     });
 
@@ -53,8 +61,8 @@ void main() {
   group('error cases', () {
     setUp(() {
       mockEntityRepository = MockEntityRepository();
-      usecase = GetAllLocations(mockEntityRepository);
-      when(() => mockEntityRepository.getAllLocations(true))
+      usecase = GetAssetLocations(mockEntityRepository);
+      when(() => mockEntityRepository.getAssetLocations())
           .thenAnswer((_) async => const Err(ServerFailure('oh no!')));
     });
 
