@@ -1,11 +1,12 @@
 //
-// Copyright (c) 2023 Nathan Fiedler
+// Copyright (c) 2024 Nathan Fiedler
 //
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:tanuki/core/domain/usecases/providers.dart';
 import 'assign_attributes_bloc.dart';
 import 'bulk_update_bloc.dart';
 import 'ingest_assets_bloc.dart';
+import 'raw_locations_bloc.dart';
 import 'recent_imports_bloc.dart';
 import 'upload_file_bloc.dart';
 
@@ -22,6 +23,12 @@ final bulkUpdateBlocProvider = Provider.autoDispose<BulkUpdateBloc>(
 final ingestAssetsBlocProvider = Provider.autoDispose<IngestAssetsBloc>(
   (ref) => IngestAssetsBloc(
     usecase: ref.read(ingestAssetsUsecaseProvider),
+  ),
+);
+
+final rawLocationsBlocProvider = Provider.autoDispose<RawLocationsBloc>(
+  (ref) => RawLocationsBloc(
+    usecase: ref.read(getAssetLocationsUsecaseProvider),
   ),
 );
 
