@@ -1,57 +1,52 @@
 //
 // Copyright (c) 2024 Nathan Fiedler
 //
-use crate::domain::entities::LabeledCount;
+use crate::domain::entities::{LabeledCount, Location};
 use crate::domain::repositories::RecordRepository;
+use crate::domain::usecases::NoParams;
 use anyhow::Error;
-use std::cmp;
-use std::fmt;
 
-pub struct AllLocations {
+///
+/// Returns the various location values and their respective counts. This
+/// includes location labels that have been split on commas, the location city
+/// and region, all returned as separate elements.
+///
+pub struct PartedLocations {
     repo: Box<dyn RecordRepository>,
 }
 
-impl AllLocations {
+impl PartedLocations {
     pub fn new(repo: Box<dyn RecordRepository>) -> Self {
         Self { repo }
     }
 }
 
-impl super::UseCase<Vec<LabeledCount>, Params> for AllLocations {
-    fn call(&self, params: Params) -> Result<Vec<LabeledCount>, Error> {
-        if params.raw {
-            self.repo.raw_locations()
-        } else {
-            self.repo.all_locations()
-        }
+impl super::UseCase<Vec<LabeledCount>, NoParams> for PartedLocations {
+    fn call(&self, _params: NoParams) -> Result<Vec<LabeledCount>, Error> {
+        self.repo.all_locations()
     }
 }
 
-#[derive(Clone, Default)]
-pub struct Params {
-    /// Return the location labels as originally entered.
-    pub raw: bool,
+///
+/// Returns the unique location records with their full structure and original
+/// values.
+///
+pub struct CompleteLocations {
+    repo: Box<dyn RecordRepository>,
 }
 
-impl Params {
-    pub fn new(raw: bool) -> Self {
-        Self { raw }
+impl CompleteLocations {
+    pub fn new(repo: Box<dyn RecordRepository>) -> Self {
+        Self { repo }
     }
 }
 
-impl fmt::Display for Params {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "Params({})", self.raw)
+impl super::UseCase<Vec<Location>, NoParams> for CompleteLocations {
+    fn call(&self, _params: NoParams) -> Result<Vec<Location>, Error> {
+        self.repo.raw_locations()
     }
 }
 
-impl cmp::PartialEq for Params {
-    fn eq(&self, other: &Self) -> bool {
-        self.raw == other.raw
-    }
-}
-
-impl cmp::Eq for Params {}
 #[cfg(test)]
 mod tests {
     use super::super::UseCase;
@@ -80,9 +75,8 @@ mod tests {
         mock.expect_all_locations()
             .returning(move || Ok(expected.clone()));
         // act
-        let usecase = AllLocations::new(Box::new(mock));
-        let params = Params { raw: false };
-        let result = usecase.call(params);
+        let usecase = PartedLocations::new(Box::new(mock));
+        let result = usecase.call(NoParams {});
         // assert
         assert!(result.is_ok());
         let actual = result.unwrap();
@@ -95,36 +89,37 @@ mod tests {
     #[test]
     fn test_all_locations_raw() {
         // arrange
-        let expected = vec![
-            LabeledCount {
-                label: "hawaii".to_owned(),
-                count: 42,
+        let input = vec![
+            Location {
+                label: Some("beach".to_owned()),
+                city: Some("Waikiki".into()),
+                region: Some("Hawaii".into()),
             },
-            LabeledCount {
-                label: "paris, france".to_owned(),
-                count: 101,
+            Location {
+                label: None,
+                city: Some("Portland".into()),
+                region: Some("Oregon".into()),
             },
-            LabeledCount {
-                label: "london".to_owned(),
-                count: 14,
+            Location {
+                label: Some("black sands beach".to_owned()),
+                city: None,
+                region: None,
             },
         ];
+        let expected = input.clone();
         let mut mock = MockRecordRepository::new();
         mock.expect_raw_locations()
-            .returning(move || Ok(expected.clone()));
+            .returning(move || Ok(input.clone()));
         // act
-        let usecase = AllLocations::new(Box::new(mock));
-        let params = Params { raw: true };
-        let result = usecase.call(params);
+        let usecase = CompleteLocations::new(Box::new(mock));
+        let result = usecase.call(NoParams {});
         // assert
         assert!(result.is_ok());
         let actual = result.unwrap();
         assert_eq!(actual.len(), 3);
-        assert!(actual.iter().any(|l| l.label == "hawaii" && l.count == 42));
-        assert!(actual.iter().any(|l| l.label == "london" && l.count == 14));
-        assert!(actual
-            .iter()
-            .any(|l| l.label == "paris, france" && l.count == 101));
+        assert_eq!(actual[0], expected[0]);
+        assert_eq!(actual[1], expected[1]);
+        assert_eq!(actual[2], expected[2]);
     }
 
     #[test]
@@ -134,9 +129,8 @@ mod tests {
         mock.expect_all_locations()
             .returning(|| Err(anyhow!("oh no")));
         // act
-        let usecase = AllLocations::new(Box::new(mock));
-        let params = Params { raw: false };
-        let result = usecase.call(params);
+        let usecase = PartedLocations::new(Box::new(mock));
+        let result = usecase.call(NoParams {});
         // assert
         assert!(result.is_err());
     }
